dash-app2/requirements.txt
numpy<2
dash
dash_bootstrap_components
plotly

cufflinks
requests
gunicorn
python-dotenv
statsmodels
